+++ packages/react/build.config.ts
@@ -11,5 +11,8 @@ export default defineBuildConfig({
         }
       },
     },
+    esbuild: {
+      jsx: "automatic",
+    },
   },
 })